+++ CHANGELOG.md
@@ -5,6 +5,17 @@ details, see the commit logs. The format is based on
 [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project
 adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Changed
+
+- Default Python version to 3.10.
+- Repository default branch from master to main.
+
+### Removed
+
+- Python 3.7 support for Flake8 and Mkdocs due to dependency incompatibility.
+
 ## 0.1.0 - 2021-05-19
 
 ### Added